Cleaned up a bit by wrapping gst state functions.
The gst_element_get_state() and gst_element_set_state() functions are
now called in wrapper functions that do some kind of poor error
handling (log the error). The error handling is no worse than it was
before, but at least the logging clutter is now out of the way.