Emitting more D-Bus signals, fixed NowPlaying.
The Playback.Stopped and Playback.Paused signals are now emitted. The
emissions are done in a Gstreamer signal handler that gets called on
any element state changes. We filter for the playbin element and react
on the corresponding state changes.

The Playback.NowPlaying signal is now suppressed in case the current
stream was changed, but is not playing (this happens when Next is
called in pause mode). Instead, it is now also emitted when the stream
state changes to playing mode.