Avoid length limitation of passed URLs (#255).
There was an arbitrary limit of 512 bytes per URL, caused by buffers
of fixed size for each entry in the URL FIFO. We are now using
strdup() and thus dynamically allocated memory for URLs.

Many memory leaks were reported by Valgrind that had to be fixed as
well. (These leaks were no true leaks in the previous version because
of the URL buffers were part of the URL FIFO entries.)

Change-Id: Ib155213432878c29bcea9696ee8399adaf8a516b