Major improvement of error handling.
Large parts of the streamer core have been rewritten to get error
handling done.

The burden of error recovery/resolution is pushed to the client as
there is no way for the streamplayer to know what exactly it should do
in case of error. It all depends on application context.

Change-Id: I2e3dbc23908f388c25db3f3a247c43a1bc657bc4