Better way to suppress stop signals during skips (#280).
The problem is that there seemed to be no way to distinguish stream
stops from skips to next track because both exhibit exactly the same
state changes. Some extra bit of information must be taken into
account to make the distinction, and previously we did this with a
counter. For each skip, the counter was increased, for each stop, the
counter was decreased until it reached 0 again, finally causing the
signal emission.

Trouble is, for fast skipping, this mechanism doesn't work reliably.
GStreamer (or playbin, who knows...) starts compressing state
transitions for fast stream switches, making us lose precious
information about when to decrease our counter. The result was bad
skipping behavior and keeping in play view even on real stop events.

Instead of a counter, we are now asking playbin for the current URI it
is playing when receiving something that looks like a stop event. In
case the URI is a NULL pointer, the event is really a stop event;
otherwise, the event must be something else, i.e., a skip event.

Change-Id: Ic4ded63b760bb646a0f8c6c3b8cdcd057c814c56