Query stream duration on stream start (#563).
There is a GStreamer message about changed duration, but it comes too
early (at stream prefetch for gapless playback, not at audible stream
change). Consequently, querying the stream duration at that time
resulted in the duration of the currently playing stream, that that
was used for the next track. To fix this, we always query the stream
duration at stream start now (the corresponding GStreamer message
refers to audible stream change).

Change-Id: I47c03e5f9a25b237ee86cdfbc12d7f217f99a6b5