Avoid deadlock during error recovery (#578).
While recovering from failure, some signal handlers may be active and
wait for the streamer data lock. Tearing down the pipeline, however,
implies waiting for active signal handlers to complete, and because we
are tearing down the pipeline while holding the lock, we end up with a
deadlock.

Manifestation of this problem is certainly dependent on timing, but
may also depend on the concrete plugins involved in the pipeline at
hand.

This commit allows signal handlers to run and finish fast during error
recovery.

Change-Id: I4d96f79030e1057de7165ab0b75fd5c557f4820a