Speed up cover art transmission (#500).
TLDR: Stuttering audio in conjunction with big cover art data has been
      fixed.

There were two problems that caused significant delay of playback
under certain circumstances:

1. Cover art was sent to TACAMan synchronously. For big cover art
   data, this has caused a rather long delay (sometimes hundreds of
   milliseconds) inside our GStreamer message bus handler. Changing
   the D-Bus communication to an asynchronous method call, the delay
   is down to around 40 ms for big pictures (tested with 2 MiB).

2. For certain GStreamer pipelines, the cover art is reported several
   times (like, 20 times in a row). This happens for FLAC playback
   with embedded cover arts. Thus, the cover art data was transmitted
   as many times to TACAMan, synchronously so. We are now filtering
   the multiply reported cover art such that only a single picture is
   transmitted in most cases.

The two problems combined led to delays in the bus handler on the one
hand, thus problems with GStreamer data flow, and to high CPU load on
the other hand, further disrupting GStreamer's data flow. For certain
files it took the whole system quite some time to recover from this
stress, leading to the observed delays of start of playback in the
tens of seconds.

Change-Id: I2a384ecd378531856d3feaba845e36d36d6b9536