Streamer: Filter tags only when preparing for D-Bus.
The advantage of this is that we do not have to copy each GstTagList
for filtering when reported by GStreamer. Instead, we simply take all
the tags and merge them into the list stored inside our stream data.
When it is time to send the list over D-Bus, the filter is applied to
construct the D-Bus message.

This means we now also have all the tags ever reported available,
which may help with introspection in the future.

Change-Id: Iadb449b8342a5da243b88ea42d117688da6ff6d5