streamer: Fix queue handling at stream boundaries.
Once again, here is a fix for handling gapless playback. We need to
dequeue the stream information at the right time, i.e., when a new
stream actually starts, not when setting the URI of the follow-up
stream.

This commit fixes meta data handling for streams which contain their
own meta data, but leave some fields empty. If in this case the
preceding stream also contained meta data, then the meta data of the
next stream was merged into the meta data of the old stream, and the
merged meta data propagated to the new stream by further merges.

Change-Id: I64d9b5567570c8547d50bd0805671659115ea355