queue: Remove duplicate queue instances.
There were two queues around, one which was used by most of the code
(so the program actually worked correctly), and one which was used by
D-Bus method de.tahifi.Streamplayer.URLFIFO.Clear() (which was not in
use up until now so that the bug was not discovered before).

The Clear method had its own queue instance it was clearing when asked
to while the rest of the code worked with a different queue. The bug
was introduced in 0d1e09e4478298c308e786fc3e2e793bfc9e2f26.

Change-Id: Iaf86deca1421b56470f73124ba5296c64b942230