playback: Avoid flooding the log with "beats-per-minute" messages.
Seems like there are files which contain the "beats-per-minute" tag
(GST_TAG_BEATS_PER_MINUTE), and this tag is emitted dozens of times
for each such file while playing. It doesn't hurt much, but the log is
unnecessarily cluttered with messages telling us that the tag isn't
recognized. So we add this tag to our filter.

Seen in a log file sent by a customer.

Change-Id: I047520f1cb4e9e0f62767f97ebf9ee2ea1c115ba