streamer: Changed player settings again, adjustable via command line.
The block size is still 512 kiB, ALSA latency is 100 ms, ALSA buffer
size is 5 seconds. These are the defaults. All of these parameters can
be changed via command line parameters now.

Play/pause reaction is back to normal and the player shows very good
performance on CM1 over WLAN on the latest kernel if it is configured
for UP mode.

Change-Id: Icde57b92b83617a5dbc1c9e893ccb2e2ca08fdff