player: Allow certain non-standard tags in preset tags (#1538).
We need this to to set alttrack information when DRCP sends its
display update. The alttrack tag in the display XML should be a string
sent by the plainurl client (SPI), and so it should in case a REST API
client is in control. In the latter case, however, alttrack was set to
the URL sent by the REST API client, which seems to confuse the SPI
slave.

To fix this, we allow REST API clients to send the "x-drcpd-title"
tag, which is parroted back to drcpd when the track is playing. drcpd
recognizes this tag and treats it properly.

Change-Id: Ie0a69faa60ae96d63885b92c1f204d1c582fbebd