streamer: Immediately set next URI if about-to-finish has fired early.
New playbin3 can send the about-to-finish signal very early, sometimes
right after a new stream has started. The next URI may not have been
queued yet when it is still being retrieved from some source, so it
must be passed to the pipeline later, as soon as it becomes available.

Change-Id: I5128c1199fbe06b06ad621cd8a7ff47cc5848212