dbus: Drop support for setting playback speed.
This feature was never used, and probably never will be. Removing this
feature allows us to avoid sending D-Bus signals for communicating the
current speed factor. Such a message used to be sent around after each
seek.

The corresponding D-Bus methods return an error now, and are subject
to be removed from the D-Bus interface soon.

Change-Id: I52fe336a7a3dd7ffaf9d6ff9b2be19861aac7508