streamer: Set next uri only when GStreamer requested it.
We must set the next uri only when GStreamer is requesting it in it's
about-to-finish signal handler, or after GStreamer has requested it
(this case applies if GStreamer has requested the next uri while our
queue was still empty). Failing to do this correctly means that
gapless playback won't work with playbin3.

Change-Id: Ib16a25efe6ff32924b1dc96759360fd76848ffdf