streamer: Send more suitable error code on stream failure (#1823).
Some Internet radio streams frequently fail for no reason, and the
error code from GStreamer is the generic GST_STREAM_ERROR_FAILED code
in this case. We mapped this to StoppedReasons::Reason::WRONG_TYPE,
but this doesn't cause drcpd to restart the stream. So we are
returning StoppedReasons::Reason::PROTOCOL now.

Change-Id: I58b46d0fdb0288298b611289cb51cde5d58c27f3