usb: Fix playing from local files (#1911).
Unfortunately, we need to distinguish local from non-local streams
now. The immediate switch to buffering mode introduced in 60784919d
and 7e930c344 are required for good stream starts with hires streams,
but this prevents local files from start playing at all.

So, in this patch we generalize the stream type recognition previously
implemented for error handling such that we can use the information
much earlier.

Change-Id: I4d0d7b2fe1a04125c6a7c223a24c1225fa5472b8