streamer: Detect pipeline replacements.
This patch is supposed to make general operation more stable. Before
reacting to some signal, we check if the playbin3 pipeline has been
replaced by a new incarnation and if the signal was meant for use with
the old one. If this is case, then (most frequently) don't do
anything.

Change-Id: I7a7823fd93b1f0f8372dfcbac254bb62cfdeffea