streamer: Do not query stream duration in progress report (#1932).
A wrong duration will be reported near the end of the stream, which is
in fact the duration of the next stream. Thus, the current stream
position is almost certainly suddenly reported as either beyond the
total length or far from total length shortly before end of stream,
leading to jumping time scales on display and in app.

The stream duration is updated in other locations anyway, so simply
removing the duration query in the periodic progress report function
avoids these glitches.

Change-Id: I872805d03936fbaa56de5619586230d7fdcfd755